+++ ui/appui-react/CHANGELOG.md
@@ -1,5 +1,42 @@
 # Change Log - @itwin/appui-react
 
+## 5.26.0
+
+### Minor Changes
+
+- f291611: Deprecated `PreviewFeaturesProviderProps` interface. Use `React.ComponentProps<typeof PreviewFeaturesProvider>` instead.
+- f291611: Bumped the release tag of `PreviewFeaturesProvider` component and `PreviewFeatures` interface from `@beta` to `@public`.
+- 2c75727: Added `widgetTabActions` preview feature, which when enabled displays widget tab actions for individual tabs instead of the default behavior where only active tab actions are exposed. Note that this feature replaces the overflow dropdown menu with a scrollable container to display all tab actions. Close button will be shown in widget tab when `controlWidgetVisibility` preview feature is enabled.
+
+  ```tsx
+  <PreviewFeaturesProvider
+    features={{
+      widgetTabActions: true,
+    }}
+  >
+    <App />
+  </PreviewFeaturesProvider>
+  ```
+
+### Patch Changes
+
+- @itwin/components-react@5.26.0
+- @itwin/core-react@5.26.0
+- @itwin/imodel-components-react@5.26.0
+
+## 5.25.0
+
+### Minor Changes
+
+- 488c5cd: Add option to disable search in WidgetContentLayout
+
+### Patch Changes
+
+- Updated dependencies [091fbdc]
+  - @itwin/imodel-components-react@5.25.0
+  - @itwin/components-react@5.25.0
+  - @itwin/core-react@5.25.0
+
 ## 5.24.2
 
 ### Patch Changes